pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "ofd-global-nat"
version = "1.0.0"
description = "OFD Global NAT Data"
authors = ["Network to Code <info@networktocode.com>"]
license = "PROPRIETARY"

[tool.poetry.dependencies]
python = "^3.8"
helper-fts = "^1.0.7"
pa-pano = "^1.0.0"
cp-mgmt = "^1.0.0"
pan-os-python = "^1.6.0"
xmltodict = "^0.12.0"
pymongo = "^4.0.1"
pynautobot = "^1.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
pytest-mock = "^3.3.1"
pytest-cov = "^2.10.1"
requests_mock = "^1.7.0"
pyyaml = "^5.3"
black = "^19.10b0"
pylint = "^2.4.4"
pydocstyle = "^5.0.2"
yamllint = "^1.20.0"
bandit = "^1.6.2"
invoke = "^1.4.1"
flake8 = "^3.8.3"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]

[[tool.poetry.source]]
name = "nexus"
url = "https://nexus-appdev.1dc.com/repository/pypi-group/simple"
default = true

[[tool.poetry.source]]
name = "gitlab"
url = "https://git-enterprise-jc.onefiserv.net/api/v4/projects/1656/packages/pypi/simple"
secondary = true

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
disable = """,
    line-too-long,
    bad-continuation,
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
        /(
          \.git
        | \.tox
        | \.venv
        | env/
        | _build
        | build
        | dist
        )/
    '''

